[refactor] IconText -> TextWithIcon 컴포넌트 명 변경
- IconText보다 TextWithIcon 이라는 이름이 좀 더 의미를 파악하기 쉽고 직관적이라고 생각해 변경
