+++ client/src/components/molecules/TextWithIcon/index.ts
@@ -1,3 +1,3 @@
-export { default } from './IconText';
+export { default } from './TextWithIcon';
 
 export type AlignType = 'left' | 'right';
